pyproject.toml
[project]
name = "serverest-api-tests"
version = "0.1.0"
description = "API tests for ServeRest"
readme = "README.md"
requires-python = ">=3.11,<4.0"
authors = [
    { name = "Serhii Smetanskyi", email = "mail@serhiismetanskyi.cv" },
]
dependencies = [
    "assertpy",
    "faker",
    "pytest",
    "pytest-html",
    "python-dotenv>=1.0.0,<2.0.0",
    "requests",
    "ruff>=0.14.6",
]

[tool.ruff]
line-length = 120
target-version = "py311"

# Exclude directories
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
# Enable rules
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]

# Ignore specific rules
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # function calls in argument defaults
    "N802",  # function name should be lowercase
    "N803",  # argument name should be lowercase
    "N806",  # variable in function should be lowercase
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)s: %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
